hjlkl

+++ Pizza delivery/Models/Pizza.cs
@@ -14,12 +14,13 @@ namespace Pizza_delivery.Models
         [Required]
         public string Allergyinfo { get; set; }
         [Required]
-        public decimal Price { get; set; }
+        public double Price { get; set; }
         public string ImgUrl { get; set; }
         public string ImgThumbnailUrl { get; set; }
         public bool IsPizzaOfTheWeek { get; set; }
         public bool InStock { get; set; }
         public int CategoryId { get; set; }
+
         public Category Category { get; set; }
     }
 }
